refactor: UserDetails를 구현하는 DTO 생성
