+++ src/main/java/com/yo/webtoon/model/dto/UserDetail.java
@@ -1,0 +1,47 @@
+package com.yo.webtoon.model.dto;
+
+import com.yo.webtoon.model.constant.Role;
+import java.time.LocalDateTime;
+import java.util.Collection;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class UserDetail implements UserDetails {
+
+    private String userId;
+    private String password;
+    private LocalDateTime deleteDatetime;
+    private Role role;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(role.name()));
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public String getUsername() {
+        return userId;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return deleteDatetime == null;
+    }
+}
