+++ src/main/java/com/yo/webtoon/service/UserService.java
@@ -1,10 +1,14 @@
 package com.yo.webtoon.service;
 
+import com.yo.webtoon.exception.WebtoonException;
+import com.yo.webtoon.model.constant.ErrorCode;
+import com.yo.webtoon.model.dto.UserDetail;
 import com.yo.webtoon.model.dto.UserDto;
 import com.yo.webtoon.model.dto.UserDto.Authorization;
 import com.yo.webtoon.model.entity.UserEntity;
 import com.yo.webtoon.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -18,9 +22,17 @@ public class UserService implements UserDetailsService {
     private final PasswordEncoder passwordEncoder;
 
     @Override
-    public UserEntity loadUserByUsername(String userId) throws UsernameNotFoundException {
-        return userRepository.findByUserId(userId)
-            .orElseThrow(() -> new UsernameNotFoundException("User ID Not found -> " + userId));
+    public UserDetail loadUserByUsername(String userId) throws UsernameNotFoundException {
+        UserEntity userEntity = userRepository.findByUserId(userId)
+            .orElseThrow(
+                () -> new WebtoonException(ErrorCode.USER_NOT_FOUND, HttpStatus.BAD_REQUEST));
+
+        return UserDetail.builder()
+            .userId(userEntity.getUserId())
+            .password(userEntity.getPassword())
+            .deleteDatetime(userEntity.getDeleteDatetime())
+            .role(userEntity.getRole())
+            .build();
     }
 
     /**
@@ -28,7 +40,7 @@ public class UserService implements UserDetailsService {
      */
     public void signUp(UserDto.SignUp signUpInfo) {
         if (userRepository.existsByUserId(signUpInfo.getUserId())) {
-            throw new RuntimeException("already exists ID -> " + signUpInfo.getUserId());
+            throw new WebtoonException(ErrorCode.ALREADY_EXIST_USER_ID, HttpStatus.BAD_REQUEST);
         }
 
         UserEntity userEntity = UserEntity.toEntity(signUpInfo);
@@ -42,7 +54,8 @@ public class UserService implements UserDetailsService {
      */
     public UserDto.Authorization authenticate(UserDto.Login loginInfo) {
         UserEntity userEntity = userRepository.findByUserId(loginInfo.getUserId())
-            .orElseThrow(() -> new RuntimeException("failed login"));
+            .orElseThrow(
+                () -> new WebtoonException(ErrorCode.FAILED_LOGIN, HttpStatus.BAD_REQUEST));
 
         if (!passwordEncoder.matches(loginInfo.getPassword(), userEntity.getPassword())) {
             throw new RuntimeException("failed login");
